% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_obj_Stan.R
\name{DM_obj_Stan}
\alias{DM_obj_Stan}
\title{Return the DMRegression Stan object constructed on the specific dataset}
\usage{
DM_obj_Stan(Y, X, iterations, chains)
}
\arguments{
\item{Y}{Matrix of \verb{n x J} responses.}

\item{X}{Matrix of \verb{n x (P+1)} covariates.}
}
\value{
The Stan object constructed on the specific dataset
}
\description{
Return the DMRegression Stan object constructed on the specific dataset
}
