% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_obj_TMB.R
\name{DM_obj_TMB}
\alias{DM_obj_TMB}
\title{Return the DMRegression TMB object constructed on the specific dataset}
\usage{
DM_obj_TMB(Y, X, beta)
}
\arguments{
\item{Y}{Matrix of \verb{n x J} responses.}

\item{X}{Matrix of \verb{n x (P+1)} covariates.}
}
\value{
The TMB object constructed on the specific dataset
}
\description{
Return the DMRegression TMB object constructed on the specific dataset
}
