% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRegression-package.R
\docType{package}
\name{DMRegression-package}
\alias{DMRegression}
\alias{DMRegression-package}
\title{The 'DMRegression' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
\author{
\strong{Maintainer}: Martin Lysy \email{mlysy@uwaterloo.ca}

}
